Add files via upload
junglelang.py is a very short translator code I made that translates words to a sort of language we used to play around with back in secondary school/college, called 'jungle language'. It basically replaces every vowel in a word with a number, starting from 1 to 5. (i.e. a=1, e=2, i=3, o=4, u=5). Also, every consonant has an "a" attached to it. For instance, my name John translates to Ja4hana; London translates to La4nada4na, etc.